Added: Moved results string to config class
